# Changes guest index back to UID