--- a/scheduler/plugin/export_test.go
+++ b/scheduler/plugin/export_test.go
@@ -18,6 +18,8 @@ package wasm
 
 import (
 	wazeroapi "github.com/tetratelabs/wazero/api"
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/uuid"
 	"k8s.io/kubernetes/pkg/scheduler/framework"
 )
 
@@ -28,7 +30,7 @@ func NewTestWasmPlugin(p framework.Plugin) *WasmPlugin {
 }
 
 func (w *WasmPlugin) SetGlobals(globals map[string]int32) {
-	if err := w.pool.doWithSchedulingGuest(ctx, 0, func(g *guest) {
+	if err := w.pool.doWithSchedulingGuest(ctx, uuid.NewUUID(), func(g *guest) {
 		// Use test conventions to set a global used to test value range.
 		for n, v := range globals {
 			g.guest.ExportedGlobal(n + "_global").(wazeroapi.MutableGlobal).Set(uint64(v))
@@ -42,11 +44,11 @@ func (w *WasmPlugin) ClearGuestModule() {
 	w.guestModule = nil
 }
 
-func (w *WasmPlugin) GetSchedulingCycleID() uint32 {
-	return w.pool.schedulingCycleID
+func (w *WasmPlugin) GetScheduledPodUID() types.UID {
+	return w.pool.scheduledPodUID
 }
 
-func (w *WasmPlugin) GetBindingCycles() map[uint32]*guest {
+func (w *WasmPlugin) GetBindingCycles() map[types.UID]*guest {
 	return w.pool.binding
 }