@sanposhiho 2023-06-16
@codefromthecrypt can you check the CI fail? I'm not sure how go.mod is related to this PR's change though.