@codefromthecrypt 2023-06-17
go.mod was related. I should have run `make check` first.